coproc: Do not reset offsets during failure recovery
- Before this fix, when the failure recovery is invoked, all sources are removed
from the pacemaker. This removes all in-memory offsets, and when the same
scripts are re-loaded we will NOT read the offsets from disk because that only
occurs at program startup (pacemaker::start()).

- The effect is having all scripts offsets set to 0 again, all materialized
topics will then be out of sync with respect to their input topics.

- Solution is pacemaker::reset(), ensures that the last read offsets exist in
memory before new scripts are ready to be started.