coproc: Fixing lower_term id bug
- In the case where multiple scripts write to the same materialized topic,
during a recovery from a crash assertions from the storage layer have been
observed, noting that the term id of a record batch is lower then the current
logs tip.

- Solution is to store the known term_id and verify that all records being
pushed to a materialized log have a term at least as large as the last known.

- If the check fails, the write fails and the offset will not be committed.
Retry policy will kick in and the write will occur again with a newer term.